test: move test-utils to test/utils.js (#488)
It's pure test-helper code, not wired via .cdsrc.json like the in-memory
exporters (which must stay in test/bookshop/lib), so it belongs at the test
root. Update the 7 importers to require('./utils'). Behavior unchanged.
