+++ test/metrics-outbox-multitenant.test.js
@@ -4,7 +4,7 @@ const { setTimeout: wait } = require('node:timers/promises')
 // Exported metric data is captured in-memory by MyInMemoryMetricReader (wired via the
 // metrics-outbox profile in .cdsrc.json) instead of scraping ConsoleMetricExporter's console.dir.
 const { latestDataPointValue, forceFlush, reset } = require('./bookshop/lib/MyInMemoryMetricReader')
-const { makeExpectEventually } = require('./bookshop/lib/test-utils')
+const { makeExpectEventually } = require('./utils')
 
 const { expect, GET, axios } = cds.test(
   __dirname + '/bookshop',
