+++ test/metrics.test.js
@@ -6,7 +6,7 @@
 const cds = require('@sap/cds')
 
 const { captured, forceFlush, reset } = require('./bookshop/lib/MyInMemoryMetricReader')
-const { makeExpectEventually } = require('./bookshop/lib/test-utils')
+const { makeExpectEventually } = require('./utils')
 
 const { expect, GET } = cds.test(__dirname + '/bookshop', '--profile', 'metrics')
